multi: Check tx inputs auto revocations flag.
This adds an isAutoRevocationsEnabled boolean to
checkTicketRedeemerCommitments and updates all callers accordingly.
This is necessary because additional validation rules will be added in
subsequent commits based on whether the automatic ticket revocations
agenda is active.

Note that this is intentionally not using blockchain.AgendaFlags, and
is instead adding an additional boolean parameter, since we should be
able to retroactively update the transaction checks to be based on just
the transaction version if the agenda passes and there are no legacy
violations (i.e., version 2 revocation transactions prior to the agenda
passing), in which case this boolean parameter could be removed.