blockchain: Add checkTicketRedeemers.
This combines the checkAllowedVotes and checkAllowedRevocations
methods into a checkTicketRedeemers function.  This reduces the number
of iterations through the stake transactions and subsequently reduces
the number of checks to determine the stake transaction type.

Further, additional validation rules for the votes and revocations in a
block will be added to this method for the upcoming automatic ticket
revocations agenda.