blockchain: Add NextExpiringTickets to BestState.
This adds a NextExpiringTickets field to the BestState type, which
represents the tickets that will expire in the next block.  This allows
consumers, such as mining template generation, to determine the tickets
that are about to expire when validating or constructing a new block.

This will be used for the automatic ticket revocations agenda since one
of the new rules for that agenda is that blocks must contain revocations
for expired tickets in the block that they become expired.