multi: Add FetchUtxoEntry to mining Config.
This adds a FetchUtxoEntry method to the mining Config.  This allows the
template generation logic to fetch unspent transaction outputs from the
point of view of the main chain tip.  This has not been necessary in the
past since typically, the mining code uses FetchUtxoView to construct a
viewpoint for transactions that it includes in a block template.
However, with the automatic ticket revocations agenda, the template
generation code will need to create new revocation transactions, which
will require fetching utxos for the input tickets.