mining: Add auto revocation priority to tx queue.
This adds a new priority for automatic revocations to the tx priority
queue.  With the automatic ticket revocations agenda, revocations are
required for all tickets that are becoming missed or expired as of the
block being created.  This update ensures that those revocations are
treated as the second highest priority after votes in the priority
queue, ahead of other non-mandatory transactions.