chaingen: Add spendable coinbase out snapshots.
This adds the ability to snapshot the current spendable coinbase output
state of the generator with a name and restore it later.  This will
allow more efficient tests to be created since they will be able to use
invalidation along with restoring the spendable outputs state to avoid
the need to go through the entire vote activation process multiple
times.