multi: Add isAutoRevocationsEnabled to CheckSSRtx.
This adds the isAutoRevocationsEnabled flag to stake.CheckSSRtx, which
in turn requires it to be added to stake.DetermineTxType and updating
all callers.

This is required for stake.CheckSSRtx since it will be updated to
return errors if revocation transactions are not valid when the
automatic ticket revocations agenda is enabled.

This is not ideal since it requires passing around the context-based
isAutoRevocationsEnabled flag in many places.

If the automatic ticket revocations agenda activates and there are no
existing version 2 revocation transactions in blocks, then this could
retroactively be removed and the updated checks can be based on the
transaction version instead.