mempool: Accept test mungers for create tickets.
This modifies CreateTicketPurchase to accept a spendable output (instead
of a transaction) as well as munge functions similar to how
CreateSignedTx works and then introduces a new method named
CreateTicketPurchaseFromTx which keeps the existing signature and is
defined in terms of the aforementioned.

This allows the test code to more easily mutate the transactions as
desired prior to them being signed which provides greater flexibility.