blockchain: Implement reject new tx vers vote.
This implements the agenda for voting on rejecting new transaction
versions until they have explicitly been enabled by a consensus rule
change as defined in DCP0008 along with consensus tests.

In particular, once the vote has passed and is active, transaction
versions greater than 3 will no longer be accepted.

This does not include the script version portion of the DCP as that will
be done in a future commit.  This implies that it is important for both
this commit and the aforementioned future commit to be merged at the
same time.

Also note that this does not implement block version bumps in the mining
and validation code that will ultimately be needed since another
consensus vote is expected to land before voting begins.

The following is an overview of the changes:

- Introduce a convenience function for determining if the vote passed
  and is now active
- Modify block validation to enforce the rejection of newer transaction
  versions in accordance with the state of the vote
- Add tests for determining if the agenda is active for both mainnet and
  testnet
- Add tests to ensure proper behavior for new transaction versions as follows:
  - New regular and stake transaction versions are rejected once the
    agenda is active
  - Existing regular and stake outputs created with newer tx versions
    are spendable before the agenda is active and remain spendable after
    it is active