secp256k1: Prepare v4.0.0.
This updates the secp256k1 copyright year in the files modified since
the previous release and serves as a base for dcrec/secp256k1/v4.0.0.

There are no updated direct dependencies in this commit.

The following direct dependencies are no longer required as compared to
the previous dcrec/secp256k1/v3.0.0 release:

- github.com/decred/dcrd/chaincfg/chainhash