dcrjson: Introduce v4 module.
The current latest code for dcrjson already has breaking changes
compared to the last release and was unfortunately not preceded by
starting a new major version module per the standard process.

Fortunately, the two modules in this repo that use dcrjson, namely
rpcclient and rpc/json/rpctypes, use any of its types directly in their
public API nor have any replacements to the code that contains the
changes that require a new major version, so neither require a major
version bump as a result.

Therefore, this is able to mostly follow the process for introducing
major API breaks as follows:

- Bump the major version in the go.mod of the affected module if not
  already done since the last release tag
- Add a replacement to the go.mod in the main module if not already done
  since the last release tag
- Update all imports in the repo to use the new major version as
  necessary
- Make necessary modifications to allow all other modules to use the new
  version in the same commit