dcrjson: Prepare v4.0.0.
This updates the dcrjson module dependencies, the copyright year in the
files modified since the previous release and serves as a base for
dcrjson/v4.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.3

The full list of updated direct dependencies since the previous
dcrjson/v3.1.0 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.3

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.