blockchain: workSorterLess -> betterCandidate.
This renames the workSorterLess function to betterCandidate and reverses
the semantics accordingly.

It was originally named workSorterLess because it doubled as the less
function for a sorting implementation, but the sorting component was
removed in favor of just finding the best candidate.  Also, the function
considers more than cumulative work for making the determination
(although cumulative work is the most important and primary criteria),
so it makes more sense to rename the function to more closely match its
intended purpose.