multi: Flush cache before fetching UTXO stats.
This updates the UTXO cache to flush all entries to the backend before
UTXO stats are fetched.  This ensures that the UTXO stats are always in
sync with the current best block.

This also does not impact the effectiveness of the UTXO cache since
entries are still only evicted if necessary.