blockchain: Fix ticket db disconnect revocations.
This addresses an issue in the ticket database when disconnecting a
block that contains an automatic revocation.

When the automatic ticket revocations agenda is enabled, a ticket can
become missed and revoked in the same block.  When connecting that block
to the ticket database, this is recorded as two separate entries in the
undo data (the first moving it to missed, and the second moving it to
revoked).  Therefore, when disconnecting the block, the undo data needs
to be applied in reverse order.