stake: Correct treasury enabled vote discovery.
This modifies FindSpentTicketsInBlock to also properly detect votes that
include information to vote on treasury spends as intended.

It should be noted that this commit only corrects the function and the
issue does not affect consensus in a way that could lead to an
unintended fork.

However, the incorrect behavior does mean the stake ticket database and
block index now have invalid entries in them that will be corrected via
a database migration in a separate commit.

More specifically, the behavior has resulted in the ticket database
incorrectly marking the associated tickets as missed instead of voted,
but that information is only used for serving RPC data.  The live ticket
pool is still accurate and all votes and payouts are unaffected.

Further, the associated data in the block index is only used when
tallying vote information and since an upgrade is required before a new
vote can start anyway, it will necessarily be resolved prior to that
point.