blockchain: Correct treasury spend vote data.
This adds code to correct the ticket database and block index treasury
spend vote data due to the issue in FindSpentTicketsInBlock described by
the previous commit.

It involves unwinding all of the blocks from the current tip back to the
known good point in the ticket database and replaying them to apply the
the correct information to the various buckets in the ticket database as
well as the relevant entries in the block index.  The process can be
interrupted at any point and future invocations will resume from the
point it was interrupted.

Also, since the next release of the software will include a vote to
change the consensus rules, all of the blocks previously marked as
having failed validation need to be unmarked so they are eligible for
which ensures clients that did not update prior to new rules activating
are able to automatically recover under the new rules without having to
download the entire chain again.

Since both cases require a database version bump, this takes advantage
of the database version bump to do both a single version bump.