uint256: Introduce package infrastructure.
Profiling the CPU usage during an initial chain sync shows that roughly
65-70% of all time is spent in garbage collection operations.  This is
primarily the result of a large number of in-use allocations.  Profiling
the in-use allocations shows that around almost a quarter of all in-use
allocations (~22%) are due to standard library big integers which
require allocations.  In other words, eliminating those allocations
should lead to a speedup of around 10% to the initial chain sync.

More specifically, the allocations in question are the result of several
important calculations which could be done without allocations, and more
efficiently in terms of execution time, via fixed precision unsigned
256-bit integers.

Thus, motivated by the previous discussion, this is part of a series of
commits that implements highly optimized allocation free fixed precision
unsigned 256-bit integer arithmetic that can ultimately be used in place
of the standard library big integers.

For the time being, the package is introduced into the internal staging
area for initial review.

The following is a brief overview of the main features and benefits:

- Strong focus on performance and correctness
- Every operation is faster than the stdlib big.Int equivalent and most
  operations, including the primary math operations, are significantly
  faster
- Allocation free
  - All non-formatting operations with the specialized type are allocation free
- Supports boolean comparison, bitwise logic, and bitwise shift operations
- All operations are performed modulo 2^256
- Ergonomic API with unary-style arguments as well as some binary variants
- Conversion-free support for interoperation with native uint64 integers
- Direct conversion to and from little and big endian byte arrays
- Full support for formatted output and common base conversions
  - Formatted output uses fewer allocations than stdlib big.Int
- 100% test coverage
- Comprehensive benchmarks

In order to help ease the review process, the full implementation will
be done across many commits.

This commit only contains the basic type definition and ability to set
it to a uint64 or another uint256 along with tests, so it is not very
useful on its own.

Future commits will implement support for interpreting and producing big
and little endian bytes, the primary arithmetic operations (addition,
subtraction, multiplication, squaring, division, negation), bitwise
operations (lsh, rsh, not, or, and, xor), comparison operations (equals,
less, greater, cmp), and other convenience methods such as determining
the minimum number of bits required to represent the current value,
whether or not the value can be represented as a uint64 without loss of
precision, and text formatting with base conversion.