uint256: Add general comparison support.
This adds the ability to compare a uint256 to another uint256 or a
uint64 along with associated tests to ensure proper functionality.

The comparison logic is the usual logic where -1 is returned when the
uint256 is less than the value being compared to, 0 when it is equal,
and 1 when it is greater.

This is part of a series of commits to fully implement the uint256
package.