uint256: Add addition support.
This adds support for uint256 addition (modulo 2^256) along with
associated tests to ensure proper functionality.

It includes adding a uint256 to an existing one (a += b), adding two
uint256s and assigning the result to a third (a = b + c), and adding a
uint64 to an existing uint256 (a += uint256(b)).

This is part of a series of commits to fully implement the uint256
package.