uint256: Add squaring support.
This adds support for uint256 squaring (modulo 2^256) along with
associated tests to ensure proper functionality.

It includes squaring an existing uint256 (a *= a) and assigning the
result of squaring a uint256 to a second one (a = b*b).

This is part of a series of commits to fully implement the uint256
package.