uint256: Add division support.
This adds support for uint256 division (modulo 2^256) along with
associated tests to ensure proper functionality.

It includes dividing an existing uint256 by a second one (a /= b),
dividing two uint256s and assigning the result to a third (a = b/c), and
dividing an existing uint256 by a uint64 (a /= uint256(b)).

This is part of a series of commits to fully implement the uint256
package.