uint256: Add negation support.
This adds support for uint256 negation (modulo 2^256) along with
associated tests to ensure proper functionality.

It includes negating an existing uint256 (a = -a) and assigning the
result of negating a uint256 to a second one (a = -b).

This is part of a series of commits to fully implement the uint256
package.