uint256: Add bitwise left shift support.
This adds support for uint256 bitwise left shifting along with
associated tests to ensure proper functionality.

It includes left shifting an existing uint256 (a << b) and assigning the
result of left shifting a uint256 to a second one (a <<= b).

This is part of a series of commits to fully implement the uint256
package.