uint256: Add bitwise right shift support.
This adds support for uint256 bitwise right shifting along with
associated tests to ensure proper functionality.

It includes right shifting an existing uint256 (a >> b) and assigning
the result of right shifting a uint256 to a second one (a >>= b).

This is part of a series of commits to fully implement the uint256
package.