uint256: Add conversion to stdlib big int support.
This adds convenience methods for converting a uint256 to a standard
library big integer along with associated tests to ensure proper
functionality.

It includes a method that allows an existing big integer to be reused
thereby potentially saving allocations as well as a method that returns
a new big integer.  The latter is often more convenient to use, but is
also virtually guaranteed to cause an allocation.

This is part of a series of commits to fully implement the uint256
package.