primitives: Add difficulty bits <-> uint256.
This implements code for converting between difficulty bits and the new
uint256 type along with associated tests.

The functions are the semantic equivalents of CompactToBig and
BigToCompact from blockchain/standalone updated to use the new uint256
type instead of stdlib big integers and renamed to DiffBitsToUint256 and
Uint256ToDiffBits, respectively.

The primary motivation for renaming the functions to specifically call
out difficulty bits instead of the shorter "Compact" is to prevent
future ambiguity with Satoshi-style variable length compact integers
that will ultimately also be available from the package.

Note that for legacy reasons, even though the difficulty bits encoding
is only used in Decred to encode unsigned 256-bit integers which
represent difficulty targets, the encoding is capable of representing
negative numbers as well as numbers much larger than the maximum value
of an unsigned 256-bit integer.  Therefore, in order to enable proper
error detection and stay consistent with legacy code, DiffBitsToUint256
returns additional flags to indicate the aforementioned conditions.