wire: Write message header directly.
This modifies the message writing code that writes out message headers
to write the fields directly instead of creating a copy of the data
first.

The primary motivation for this change is that the hdr.command field is
assigned and never used in the existing code because the size-limited
array variant of the command field is used instead when writing the
elements.  Since that field is unused, rather than creating a local
partially-filled message header, and copying data into it only to turn
around and write that data immediately, it makes more sense to just
write the data directly.