blockchain: UTXO database migration fix.
This fixes an edge case within the UTXO database migration logic.  In
the upgrade that migrates the UTXO data to use simple key prefixes
rather than buckets, old entries will still be copied to the new
location even if they are nil, which incorrectly creates an empty entry
in the new location.