chaincfg: Prepare v3.1.0.
This updates the chaincfg module dependencies and serves as a base for
chaincfg/v3.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.3
- github.com/decred/dcrd/wire@v1.5.0

The full list of updated direct dependencies since the previous
chaincfg/v3.0.0 release are the same as above.

Finally, all modules in the repository are tidied to ensure they are
updated to use the latest versions hoisted forward as a result.