multi: Use dcrutil Tx in NewTxDeepTxIns.
This updates dcrutil.NewTxDeepTxIns to take a dcrutil.Tx as a parameter
rather than a wire.MsgTx, which allows for the tree and index set on the
input dcrutil.Tx to be copied to the new dcrutil.Tx.

This is less error prone for callers since the existing implementation
is often used to replace an existing dcrutil.Tx with the deep copy that
this function creates, but a side effect that is easy to miss is that
the new dcrutil.Tx would have reset the tree and index fields which
can lead to unintended behavior.