rpc/jsonrpc/types: Prepare v3.0.0.
This updates the rpc/jsonrpc/types module dependencies, the copyright
year in the files modified since the previous release, and serves as a
base for rpc/jsonrpc/types/v3.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/dcrjson/v4@v4.0.0

The full list of updated direct dependencies since the previous
rpc/jsonrpc/types/v2.3.0 release are the same as above.