mempool: Update fraud proof data.
This updates the fraud proof data on transaction inputs as necessary
when entering the mempool.  The fraud proof data will ultimately get
filled in properly when miners create a block template.  However,
ensuring that it is filled in properly when entering the mempool is
beneficial so that it is correct when relaying the transaction or
returning it from an RPC API.