stdscript: Add script to address conversion.
As part of the ongoing effort to refactor all code dealing with standard
scripts out of txscript, this adds support to stdscript for extracting
the addresses from standard recognized forms of public key scripts along
with comprehensive tests to ensure proper functionality.

Staying consistent with the design of determining script types, the API
is designed such that there is an exported function specifically for
extracting addresses from version 0 scripts as well as a separate
variant that accepts the script version.