txscript: Remove unused ExtractPkScriptAltSigType.
This removes ExtractPkScriptAltSigType and its associated benchmark
since it is no longer used by anything in the repository and is now
available via stdscript.ExtractPubKeyAltDetailsV0 or
stdscript.ExtractPubKeyHashAltDetailsV0 although it typically is no
longer needed since stdscript has more fine-grained script types.

This is part of a series of commits to remove all code related to
standard scripts from txscript.