txscript: Remove unused GetStakeOutSubclass.
This removes GetScriptClass since it is no longer used by anything in
the repository and is also no longer needed since the relevant
information is now available via the more fine grained script types
of stdscript.DetermineScriptType.

This is part of a series of commits to remove all code related to
standard scripts from txscript.