netsync: Consolidate initial sync handling.
In an effort to consolidate the handling when the manager determines the
initial chain sync process completes, this introduces a new method named
onInitialChainSyncDone that is invoked accordingly and moves the
duplicate log statement into the new method.

This will also allow future commits to more easily add additional
handling when the initial chain sync process completes.