netsync: Request init state immediately upon sync.
This modifies the logic that requests the initial state to happen
immediately when the initial chain sync process is complete as opposed
to the existing approach that uses a separate goroutine with polling on
a 3 second interval.

This approach is more efficient since it avoids extra goroutines and
also it also allows obtaining initial data, most notably pending
treasury spend transaction and votes for the current best block, more
quickly.