hdkeychain: Prepare v3.0.1.
This updates the hdkeychain module dependencies and serves as a base for
hdkeychain/v3.0.1.

The full list of updated direct dependencies since the previous
hdkeychain/v3.0.0 release are as follows:

- github.com/decred/dcrd/chaincfg/v3@v3.1.0
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.0.1
- github.com/decred/dcrd/txscript/v4@v4.0.0

The following direct dependencies are no longer required as compared to
the previous hdkeychain/v3.0.0 release:

- github.com/decred/dcrd/dcrec

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.