dcrutil: Prepare v4.0.0.
This updates the dcrutil module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
dcrutil/v4.0.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/txscript/v4@v4.0.0

The full list of updated direct dependencies since the previous
dcrutil/v3.0.0 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.3
- github.com/decred/dcrd/dcrec/edwards/v2@v2.0.2
- github.com/decred/dcrd/dcrec/secp256k1/v4@v4.0.1
- github.com/decred/dcrd/txscript/v4@v4.0.0
- github.com/decred/dcrd/wire@v1.5.0

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.