connmgr: Prepare v3.1.0.
This updates the connmgr module dependencies, the copyright year in the
files modified since the previous release, and serves as a base for
connmgr/v3.1.0.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/wire@v1.5.0
- github.com/decred/slog@v1.2.0

The full list of updated direct dependencies since the previous
connmgr/v3.0.0 release are the same as above.

Finally, all modules in the repository that depend on the module are
tidied to ensure they are updated to use the latest versions hoisted
forward as a result.