primitives: Add core merkle tree root calcs.
This implements functions to calculate a Merkle tree root for a given
set of leaves along with associated tests.

The functions are direct ports of the functions of the same names from
blockchain/standalone.

Note that this does not include the other functions in
blockchain/standalone related to merkle root calculations from
transaction trees since the primitives package does not yet implement
a transaction type.  Those will be implemented later after a transaction
type is introduced.