rpcserver: Submit transactions directly.
This modifies the adaptor that the RPC server uses for transaction
submission to process the transaction directly instead of piping through
the netsync package.  There is no longer any reason to incur the extra
overhead since the mempool is concurrent safe the initial reasons for
piping it through netsync no longer apply.  This can be seen by noting
that the net sync manager doesn't do anything except call the mempool
method.