chaincfg: Add AssumeValid param.
This introduces an AssumeValid parameter that is the hash of a block
that has been externally verified to be valid.  In a subsequent commit,
it will be used to allow several validation checks to be skipped for
blocks that are both an ancestor of the assumed valid block and an
ancestor of the best header.

Note that the AssumeValid block hash that is specified for mainnet and
testnet is for the same block that was used for MinKnownChainWork.
Using the same block for both of these slightly simplifies the process
for external verification of the values when they are updated
periodically with new releases.