blockchain: Set BFFastAdd based on assume valid.
This updates BlockChain to set the BFFastAdd flag if a block is both an
ancestor of the assumed valid block and an ancestor of the best header.

A helper function isAssumeValidAncestor is introduced that determines
whether the provided node is both an ancestor of the assumed valid node
and an ancestor of the best header.  False is returned when the assume
valid node has not been discovered or when assume valid is disabled.

The assumed valid node is clamped back as needed to ensure that it is at
least 2 weeks worth of blocks behind the best header.  This protects
against malicious actors trying to play games by tricking users into
using an assumed valid hash near the tip.